simplify code with provider interface pattern
This will make mocking so much easier for all actions
